Decode OfferTask event from transaction receipt instead of waiting
